// src/app/interfaces/summary.interface.ts
export interface Isummary {
    ActivityCount: string;
    CharacteristicName: string;
    CharacteristicType: string;
    CountyName: string;
    HUCEightDigitCode: string;
    LastResultSubmittedDate: string;
    MonitoringLocationIdentifier: string;
    MonitoringLocationLatitude: string;
    MonitoringLocationLongitude: string;
    MonitoringLocationName: string;
    MonitoringLocationTypeName: string;
    MonitoringLocationUrl: string;
    OrganizationFormalName: string;
    OrganizationIdentifier: string;
    Provider: string;
    ResolvedMonitoringLocationTypeName: string;
    ResultCount: string;
    StateName: string;
    YearSummarized: string;
}