// src/app/interfaces/org.interface.ts
export interface Iorg {
    ElectronicAddress: string;
    OrganizationAddress: Array<string>;
    OrganizationDescriptionText: string;
    OrganizationFormalName: string;
    OrganizationIdentifier: string;
    OrganizationType: string;
    ProviderName: string;
    Telephonic: string;
    TribalCode: string;
}